cloud-functions/daily-diagnostics-logger/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
requests==2.*
